236 - updated CSS to fix list
I spent a while trying to re-build the list because it doesn't make semantic sense and also to get better control over the elements but there's a load of references and it would end up looking slightly different so i just made a quick fix. I'm not happy with the height of the text. There's some wrapping issues with the French translation but to fix that would mean changing the design for all the languages and I'm not sure on the importance / amount of traffic in FR to make this worthwhile.

In the end I went back to my original fix and changed 1 character to make the column wider.

(cherry picked from commit 00884ddb2214fa0700d3767e5e6cc7271755cebe)