Fix usage of culture dependent values to do the calculations
The values used for calculations are now culture independent.
This is a shortcut for a possible final solution, because is a bug present in production.
Is necessary to move some values to data attributes and then is possible to display the values to the user with the user culture and to use the culture independent to the calculations.

Some To-Do were added to the file.